Java 8 modernization and cleanup in all packages
Removed final on static methods
Removed final on enum declarations
Remove unused imports
Removed methods that only call superclass method
Removed redundant access modifiers on enum constructor
Made as many fields possible final
Used lambda expressions where possible
Used Map::computeIfAbsent in place of verbose construct
Stopped creating redundant arrays in log statements
Removed unnecessary boxing and unboxing
Replaced .stream().forEach() with .forEach()
Used forEach loops in place of traditional iteration

Addresses #781 in part

+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/tinkerpop/ElementUtils.java
@@ -41,7 +41,7 @@ public class ElementUtils {
             if (id instanceof Vertex)
                 return Long.parseLong(((Vertex) id).id().toString());
             else
-                return Long.valueOf(id.toString()).longValue();
+                return Long.valueOf(id.toString());
         } catch (NumberFormatException e) {
             return 0;
         }
